@@ -7,11 +7,13 @@
 		"@testing-library/react": "^11.1.0",
 		"@testing-library/user-event": "^12.1.10",
 		"prop-types": "^15.8.1",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-router": "^6.4.1",
+		"react-router-dom": "^6.4.1",
 		"react-scripts": "^5.0.1",
-		"web-vitals": "^1.0.1",
-		"sass": "^1.38.0"
+		"sass": "^1.38.0",
+		"web-vitals": "^1.0.1"
 	},
 	"scripts": {
 		"start": "react-scripts start",
@@ -36,6 +38,5 @@
 			"last 1 firefox version",
 			"last 1 safari version"
 		]
-	},
-	"devDependencies": {}
+	}
 }
