@@ -1,44 +1,23 @@
-import { useState } from "react/cjs/react.development";
 import AppHeader from "../appHeader/AppHeader";
-import RandomChar from "../randomChar/RandomChar";
-import CharList from "../charList/CharList";
-import CharInfo from "../charInfo/CharInfo";
-import ErrorBoundary from "../errorBoundary/ErrorBoundary";
-import ComicsList from	"../comicsList/ComicsList";
-import SingleComic from "../singleComic/SingleComic";
+import { MainPage, ComicsPage } from '../pages';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 
 
-import decoration from '../../resources/img/vision.png';
 
 const App = () => {
-
-	const [selectedChar, setChar] = useState(null);
-
-	const onCharSelected = (id) => {
-		setChar(id)
-	}
-
 	return (
-		<div className="app">
-			<AppHeader />
-			<main>
-				<RandomChar
-					onCharSelected={onCharSelected}
-				/>
-				<div className="char__content">
-					<CharList
-						selectedChar={selectedChar}
-						onCharSelected={onCharSelected} />
-					<ErrorBoundary>
-						<CharInfo charId={selectedChar} />
-					</ErrorBoundary>
-				</div>
-				<br/><br/><br/>
-				<ComicsList/><br/><br/><br/>
-				<SingleComic/>
-				<img className="bg-decoration" src={decoration} alt="vision" />
-			</main>
-		</div>
+		<Router>
+			<div className="app">
+				<AppHeader />
+				<main>
+					<Routes>
+						<Route path="/" element={<MainPage />} />
+						<Route path="/comics" element={<ComicsPage />} />
+					</Routes>
+				</main>
+			</div>
+		</Router>
+
 	)
 }
 
