@@ -1,20 +1,58 @@
 import './singleComic.scss';
 import xMen from '../../resources/img/x-men.png';
+import { useState, useEffect } from "react";
+import useMarvelService from '../../services/MarvelService';
+import Spinner from '../spiner/Spiner';
+import ErrorMessage from '../errorMessage/ErrorMessage';
 
 const SingleComic = () => {
-    return (
-        <div className="single-comic">
-            <img src={xMen} alt="x-men" className="single-comic__img"/>
-            <div className="single-comic__info">
-                <h2 className="single-comic__name">X-Men: Days of Future Past</h2>
-                <p className="single-comic__descr">Re-live the legendary first journey into the dystopian future of 2013 - where Sentinels stalk the Earth, and the X-Men are humanity's only hope...until they die! Also featuring the first appearance of Alpha Flight, the return of the Wendigo, the history of the X-Men from Cyclops himself...and a demon for Christmas!?</p>
-                <p className="single-comic__descr">144 pages</p>
-                <p className="single-comic__descr">Language: en-us</p>
-                <div className="single-comic__price">9.99$</div>
-            </div>
-            <a href="#" className="single-comic__back">Back to all</a>
-        </div>
-    )
+
+	const [comic, setComic] = useState([]);
+	const { loading, error, getComicsById, clearError } = useMarvelService();
+
+	useEffect(() => updateComic(1332), []);
+
+	const updateComic = (comicId) => {
+		clearError();
+		getComicsById(comicId)
+			.then(onCharLoaded)
+	}
+
+	const onCharLoaded = (comic) => {
+		setComic(comic);
+	}
+
+	// const errorMessage = error ? <ErrorMessage /> : null;
+	// const spinner = loading ? <Spinner /> : null;
+	// const content = !(loading || error || !comic) ? <View comic={comic} /> : null
+	const errorMessage = error ? <ErrorMessage /> : null;
+	const spinner = loading ? <Spinner /> : null;
+	const content = View(comic);
+
+	return (
+		<div className="single-comic">
+			{errorMessage}
+			{spinner}
+			{content}
+		</div>
+	)
+}
+
+const View = (comic) => {
+	return (
+		<>
+			<img src={comic.thumbnail} alt="x-men" className="single-comic__img" />
+			<div className="single-comic__info">
+				<h2 className="single-comic__name">{comic.title}</h2>
+				<p className="single-comic__descr">{comic.disc}</p>
+				<p className="single-comic__descr">{comic.pageCount} pages</p>
+				<p className="single-comic__descr">Language: {comic.language}</p>
+				<div className="single-comic__price">{comic.prices}$</div>
+			</div>
+			<a href="#" className="single-comic__back">Back to all</a>
+		</>
+
+	)
 }
 
 export default SingleComic;
