@@ -1,12 +1,13 @@
-import { useState, useCallback } from "react/cjs/react.development";
+import { useState, useCallback } from "react";
 
 export const useHttp = () => {
 	const [loading, setLoading] = useState(false);
 	const [error, setError] = useState(null);
 
 	const request = useCallback(async (url, method = 'GET', body = null, headers = { 'Content-Type': 'application/json' }) => {
-		setLoading(true);
 		
+		setLoading(true);
+
 		try {
 			const response = await fetch(url, { method, body, headers });
 
